Move JSON into sources package
